add ProgressBarWithExperienceCount component for connecting container
